requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-telegram-bot==21.5
flask